Chore: Add clang-tidy configuration files and fix warnings
- add global and module-level clang-tidy configuration files
- integrate clang toolchain/settings in CMake and VS Code config
- fix clang-tidy warnings in app init, network, sensors, and main code
- update CI workflow and ignore rules for clang tooling

Ref #3

+++ Core/App/Network/LwIP/BSP/eth_phy_io.c
@@ -180,5 +180,5 @@ static int32_t ETH_PHY_IO_WriteReg(uint32_t DevAddr, uint32_t RegAddr, uint32_t 
   */
 static int32_t ETH_PHY_IO_GetTick(void)
 {
-    return xTaskGetTickCount() * portTICK_PERIOD_MS;
+    return (int32_t)(xTaskGetTickCount() * portTICK_PERIOD_MS);
 }
